Work around a date corner case in .xls (forced as string)
There's a case where you can place a single quote in a cell to force
the cell to be handled like a string. For excel, there was a test
that was erroneously identifying that escaped string as a float
because of a bad test to see if it's a float. This replaces that
check with a better one