Fix broken type inference from currency to float in Excelx
The type inference that Excelx#to_type does is broken for at least
one valid currency format. Check out this xlsx file:

http://scpike-drop.s3.amazonaws.com/broken.xlsx

 > Roo::Excelx.new("broken.xlsx")
 => {[1, 1]=>"1899-12-30"}

The problem is that any format code with a "d" in it that's not in
EXCEPTIONAL_FORMATS is treated as a date. The format code for this
currency has "Red" in it. Rather than hardcoding the list of formats
with "d"s in them that are really floats, can we just assume all
formats with # are floats? My interpretation of §18.8.31 in the spec
says yes.

This is not a problem in Roo::Excel, just Roo::Excelx