Missing Header Raises `Roo::HeaderRowNotFoundError`
Currently, Roo raises a generic `RuntimeError` with the message
"Couldn't find header row" when a matching header row is not found.

This behavior makes it difficult to rescue from the most specific error
desired, because `RuntimeError` is the default error class for `raise`.

Additionally, because Roo does not have its own library-level error
class from which all errors inherit, it is difficult to differentiate
Roo library errors from errors caused by the developer / user.

From the [Ruby documentation][1]:

> It is recommended that a library should have one subclass of
> `StandardError` or `RuntimeError` and have specific exception types
> inherit from it. This allows the user to rescue a generic exception
> type to catch all exceptions the library may raise even if future
> versions of the library add new exception subclasses.

This commit adds a generic `Roo::Error` class which inherits from
`StandardError` and a specific `Roo::HeaderRowNotFound` class that is
raised when a matching header row could not be found.

In the future, additional subclasses of `Roo::Error` can be added in
order to allow users to more easily rescue the correct errors.

Note: This is *not* a backwards-compatible change. Users who currently
rescue `RuntimeError` for a missing header row will now see an error
raised.

[1]: http://ruby-doc.org/core-2.2.2/Exception.html