Fixes for cells with errors like #N/A
This fixes an issue found in #269 while making the fewest breaking behavioral changes. It deals with all of the following errors: #N/A #REF! #NAME? #DIV/0! #NULL! #VALUE! #NUM!

One are of concern: this fix means Cell::Number is capable of returning a String (e.g. #N/A, #REF!, etc.) in addition to returning a Float or Integer. That might not coincide with a users expectations when using the gem.