Parse sheets in ODS files once
This improves the performance greatly for large files because instead of parsing
the sheets everytime a cell is accessed, this commit moves the parsing to the
constructor.

In my tests with a 60K line ODS spreadsheet this change resulted in running
times around ~70 seconds. Before this change the processing was not finished
after having run for 10-20 minutes. Others are reporting > 4 hours of processing
here: https://github.com/roo-rb/roo/issues/113.