Fix test_finalize test
`test_finalize` broke in ruby 2.4 because of changes in Ruby's GC. The
initial fix was to fork the process to make sure directories where being
cleaned properly.

Unfortunately, this caused a segfault when running the entire test suite.
The issue was related to where `ObjectSpace.define_finalizer` was being
called from (a class that was being extended by a module). The issue is
probably related to Nokogiri, but I wasn't able to develop a trivial
example of the error.

Creating the finalizer in the class fixed the issue in Ruby 2.4.0.