Add a merged range without content and test to ensure we don't copy nils
If you merge empty cells, the definitions for merged ranges references
the appropriate coordinates for the merge, but the cells hash does not
include a value at that coordinate. When copying the src to dst, we end
up copying `nil` into the `cells` hash which causes errors in several
other parts of parsing.